requirements.txt
streamlit
yfinance
pandas
plotly
python-dateutil
streamlit-calendar